# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Users and Roles",
        "description": "This is a Users and Roles API using Lumen and Swagger",
        "contact": {
            "email": "rhiannon.eldridge.au@gmail.com"
        },
        "license": {
            "name": "BY-NC-SA",
            "url": "https://creativecommons.org/licenses/by-nc-sa/4.0"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Localhost Server"
        }
    ],
    "paths": {
...
        "/api/roles": {
            "get": {
                "tags": [
                    "roles"
                ],
                "summary": "Get All Roles",
                "operationId": "getRoles",
                "responses": {
                    "200": {
                        "description": "Array of all Roles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Role"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "roles"
                ],
                "summary": "Create Role",
                "operationId": "createRole",
                "requestBody": {
                    "description": "Create Role object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUpdateRole"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Role Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Role"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/roles/{id}": {
            "get": {
                "tags": [
                    "roles"
                ],
                "summary": "Get Role",
                "operationId": "getRole",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The specified Role",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Role"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "roles"
                ],
                "summary": "Update Role",
                "operationId": "updateRole",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update Role object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUpdateRole"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Role Updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Role"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "roles"
                ],
                "summary": "Delete Role",
                "operationId": "deleteRole",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The Role has been deleted"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "CreateUpdateRole": {
                "title": "Create/Update Role model",
                "description": "Create/Update Role model",
                "properties": {
                    "name": {
                        "title": "Name",
                        "description": "Name of Role",
                        "format": "string"
                    }
                },
                "type": "object"
            },
...
            "Role": {
                "title": "Role",
                "description": "Role schema",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID of Role",
                        "format": "int64"
                    },
                    "name": {
                        "title": "Name",
                        "description": "Name of Role",
                        "format": "string"
                    },
                    "created_at": {
                        "title": "Created At",
                        "description": "Date Role was created",
                        "format": "string"
                    },
                    "updated_at": {
                        "title": "Updated At",
                        "description": "Date Role was last updated",
                        "format": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "description": "Your API Key",
                "name": "X-Api-Key",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "api_key": []
        }
    ],
...
}